Initial setup of useZip custom hook
